requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
strawberry-graphql